refactor(curator): split profile into cohesive modules
